Code Review and dependencies update
The workflow uses a couple of packages like spare_score and DLMUSE. The 2 packages had conflicts before, right now we will try to solve them individually(the dependencies.txt will have the same ones for both DLMUSE and spare in the future), but snakemake builds fine with python 3.12 for now.
The only thing that we have to do now is to add the nnUnet model weights directly.